+++ src/chrome_agent_bridge/__init__.py
@@ -1,5 +1,12 @@
 """Chrome Agent Bridge public interface."""
 
+from .launch_agent import LaunchAgentManager, LaunchAgentPaths
 from .manager import BridgeManager, BridgeState, DevToolsHealth
 
-__all__ = ["BridgeManager", "BridgeState", "DevToolsHealth"]
+__all__ = [
+    "BridgeManager",
+    "BridgeState",
+    "DevToolsHealth",
+    "LaunchAgentManager",
+    "LaunchAgentPaths",
+]